Redesign feed view
