+++ plugins/card-resources/src/components/CardSection.svelte
@@ -1,0 +1,44 @@
+<!--
+// Copyright © 2025 Hardcore Engineering Inc.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+-->
+
+<script lang="ts">
+</script>
+
+<!-- svelte-ignore a11y-click-events-have-key-events -->
+<!-- svelte-ignore a11y-no-static-element-interactions -->
+<div class="section" on:click|stopPropagation|preventDefault>
+  <div class="marker" />
+  <slot />
+</div>
+
+<style lang="scss">
+  .section {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    cursor: pointer;
+    user-select: none;
+    border-radius: 0.5rem;
+    width: 100%;
+  }
+
+  .marker {
+    position: absolute;
+    width: 0.25rem;
+    height: 100%;
+    background: var(--global-ui-highlight-BackgroundColor);
+    border-radius: 0;
+  }
+</style>
